speedtest/requirements.txt
requests
paho-mqtt
speedtest-cli